word와 sentence에서 동시에 사용하는 스페인어 등록 Input을 하나의 컴포넌트로 통합 및 메인 페이지 내 단어, 문장 동시 표시
