+++ components/setence-list.tsx
@@ -1,38 +1,24 @@
 'use client';
 
 import { useState, useEffect } from 'react';
-import Input from '@/components/input';
-import { enrollSentence, getSentences } from '@/service/sentence';
-import { Sentence } from '@/types';
+import { getSpanish } from '@/service/spanish';
+import { Spanish } from '@/types';
 
-export default function SentenceList() {
-  const [spanish, setSpanish] = useState('');
-  const [korean, setKorean] = useState('');
-  const [sentences, setSentences] = useState<Sentence[]>([]);
+type Props = {
+  limitNumber?: number;
+};
 
-  const onClickHandler = () => {
-    try {
-      enrollSentence(spanish, korean);
-
-      setSentences(sentences.concat({ spanish, korean }));
-      setSpanish('');
-      setKorean('');
-    } catch (err) {
-      console.log(err);
-    }
-  };
+export default function SentenceList({ limitNumber }: Props) {
+  const [sentences, setSentences] = useState<Spanish[]>([]);
 
   useEffect(() => {
-    getSentences()
+    getSpanish('sentences', limitNumber ?? 0)
       .then((sentences) => setSentences(sentences))
       .catch((err) => console.log(err));
-  }, []);
+  }, [limitNumber]);
 
   return (
     <div>
-      <Input value={spanish} setValue={setSpanish} />
-      <Input value={korean} setValue={setKorean} />
-      <button onClick={onClickHandler}>추가</button>
       {sentences &&
         sentences.map((sentences, index) => (
           <li key={index}>
