+++ def.ts
@@ -1,0 +1,1 @@
+export const MAX_QUERY_NUMBER = 100;
